add encointer-primitives crate. Move encointer-community primitives to encointer-primitives/communities
