+++ Cargo.toml
@@ -7,5 +7,6 @@ members = [
     'bazaar',
     'sybil/gate',
     'sybil/proof-issuer',
+    'primitives',
     'test-utils'
 ]
