+++ balances/Cargo.toml
@@ -11,6 +11,10 @@ codec = { package = "parity-scale-codec", version = "1.3.0", default-features = 
 num-traits = { version = "0.2.8", default-features = false }
 impl-trait-for-tuples = "0.1.3"
 
+[dependencies.encointer-primitives]
+package = "encointer-primitives"
+path = "../primitives"
+
 [dependencies.encointer-communities]
 default-features = false
 path = "../communities"
