+++ communities/src/lib.rs
@@ -31,15 +31,16 @@ use frame_support::{
     storage::{StorageMap, StorageValue},
 };
 use frame_system::ensure_signed;
-use sp_core::{RuntimeDebug, H256};
 
 use rstd::prelude::*;
 
-use codec::{Decode, Encode};
-pub use fixed::traits::{LossyFrom, LossyInto};
+use codec::Encode;
 use fixed::transcendental::{asin, cos, powi, sin, sqrt};
-use fixed::types::{I32F0, I32F32, I64F64, U0F64};
+use fixed::types::I32F32;
 use runtime_io::hashing::blake2_256;
+
+use encointer_primitives::communities::consts::*;
+use encointer_primitives::communities::*;
 
 pub trait Trait: frame_system::Trait {
     type Event: From<Event<Self>> + Into<<Self as frame_system::Trait>::Event>;
@@ -47,47 +48,6 @@ pub trait Trait: frame_system::Trait {
 
 // Logger target
 const LOG: &str = "encointer";
-
-pub type CommunityIndexType = u32;
-pub type LocationIndexType = u32;
-pub type Degree = I32F32;
-pub type Demurrage = I64F64;
-
-// Location in lat/lon. Fixpoint value in degree with 8 decimal bits and 24 fractional bits
-#[derive(Encode, Decode, Copy, Clone, PartialEq, Eq, Default, RuntimeDebug)]
-pub struct Location {
-    pub lat: Degree,
-    pub lon: Degree,
-}
-pub type CommunityIdentifier = H256;
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq, Default, RuntimeDebug)]
-pub struct CommunityPropertiesType {
-    pub name_utf8: Vec<u8>,
-    pub demurrage_per_block: Demurrage,
-}
-
-const MAX_SPEED_MPS: i32 = 83; // [m/s] max speed over ground of adversary
-const MIN_SOLAR_TRIP_TIME_S: i32 = 1; // [s] minimum adversary trip time between two locations measured in local (solar) time.
-
-const DATELINE_DISTANCE_M: u32 = 1_000_000; // meetups may not be closer to dateline (or poles) than this
-
-const NORTH_POLE: Location = Location {
-    lon: Degree::from_bits(0i64),
-    lat: Degree::from_bits(90i64 << 32),
-};
-const SOUTH_POLE: Location = Location {
-    lon: Degree::from_bits(0i64),
-    lat: Degree::from_bits(-90i64 << 32),
-};
-const DATELINE_LON: Degree = Degree::from_bits(180i64 << 32);
-
-// dec2hex(round(pi/180 * 2^64),16)
-const RADIANS_PER_DEGREE: U0F64 = U0F64::from_bits(0x0477D1A894A74E40);
-
-// dec2hex(6371000,8)
-// in meters
-const MEAN_EARTH_RADIUS: I32F0 = I32F0::from_bits(0x006136B8);
 
 decl_storage! {
     trait Store for Module<T: Trait> as EncointerCommunities {
