+++ primitives/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "encointer-primitives"
+version = "0.1.0"
+authors = ["Christian Langenbacher <christian.langenbacher91@gmail.com>"]
+edition = "2018"
+
+[dependencies.codec]
+default-features = false
+package = "parity-scale-codec"
+features = ["derive"]
+version = "1.3.0"
+
+[dependencies.fixed]
+default-features = false
+git = "https://github.com/encointer/substrate-fixed"
+tag = "v0.5.4+sub_v0.1"
+package = "substrate-fixed"
+
+
+[dependencies.sp-core]
+default-features = false
+package = "sp-core"
+git = "https://github.com/paritytech/substrate.git"
+rev = "a208da16"
