+++ primitives/src/communities.rs
@@ -1,0 +1,51 @@
+use codec::{Decode, Encode};
+use fixed::types::{I32F32, I64F64};
+use sp_core::{RuntimeDebug, H256};
+
+pub use fixed::traits::{LossyFrom, LossyInto};
+
+pub type CommunityIndexType = u32;
+pub type LocationIndexType = u32;
+pub type Degree = I32F32;
+pub type Demurrage = I64F64;
+pub type CommunityIdentifier = H256;
+
+// Location in lat/lon. Fixpoint value in degree with 8 decimal bits and 24 fractional bits
+#[derive(Encode, Decode, Copy, Clone, PartialEq, Eq, Default, RuntimeDebug)]
+pub struct Location {
+    pub lat: Degree,
+    pub lon: Degree,
+}
+
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Default, RuntimeDebug)]
+pub struct CommunityPropertiesType {
+    pub name_utf8: Vec<u8>,
+    pub demurrage_per_block: Demurrage,
+}
+
+pub mod consts {
+    use super::{Degree, Location};
+    use fixed::types::{I32F0, U0F64};
+
+    pub const MAX_SPEED_MPS: i32 = 83; // [m/s] max speed over ground of adversary
+    pub const MIN_SOLAR_TRIP_TIME_S: i32 = 1; // [s] minimum adversary trip time between two locations measured in local (solar) time.
+
+    pub const DATELINE_DISTANCE_M: u32 = 1_000_000; // meetups may not be closer to dateline (or poles) than this
+
+    pub const NORTH_POLE: Location = Location {
+        lon: Degree::from_bits(0i64),
+        lat: Degree::from_bits(90i64 << 32),
+    };
+    pub const SOUTH_POLE: Location = Location {
+        lon: Degree::from_bits(0i64),
+        lat: Degree::from_bits(-90i64 << 32),
+    };
+    pub const DATELINE_LON: Degree = Degree::from_bits(180i64 << 32);
+
+    // dec2hex(round(pi/180 * 2^64),16)
+    pub const RADIANS_PER_DEGREE: U0F64 = U0F64::from_bits(0x0477D1A894A74E40);
+
+    // dec2hex(6371000,8)
+    // in meters
+    pub const MEAN_EARTH_RADIUS: I32F0 = I32F0::from_bits(0x006136B8);
+}
